stateless factory class implemented

+++ base/VkImageBundle.cpp
@@ -1,21 +1,21 @@
 #include "VkImageBundle.h"
 
-VkImageBundle::VkImageBundle(std::vector<std::shared_ptr<VkImageBase>> _images, uint32_t bundle_size) :
+VkImageBundle::VkImageBundle(std::vector<std::shared_ptr<VkImageBase>> _images, size_t bundle_size) :
     bundle_count(bundle_size),
     images(std::move(_images))
 {}
 
-uint32_t VkImageBundle::GetBundleCount() const
+size_t VkImageBundle::GetBundleCount() const
 {
 	return bundle_count;
 }
 
-const VkImageBase & VkImageBundle::operator[](uint32_t index) const {
+const VkImageBase & VkImageBundle::operator[](size_t index) const {
 	
 	return *images[index];
 }
 
-const VkImageBase & VkImageBundle::GetOne(uint32_t index) const
+const VkImageBase & VkImageBundle::GetOne(size_t index) const
 {
 
 	return *images[index];
